requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
requests==2.31.0
notion-client==2.2.1
anthropic==0.7.8
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
tqdm>=4.65.0
openpyxl>=3.1.0
gspread>=5.12.0 